# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/media/presigned-url": {
            "get": {
                "description": "Generate a presigned GET URL for accessing uploaded media files by S3 key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Get presigned GET URL for media access",
                "parameters": [
                    {
                        "type": "string",
                        "description": "S3 object key (e.g., documents/abc123.pdf)",
                        "name": "s3_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 900,
                        "description": "TTL in seconds (default: 900, max: 86400)",
                        "name": "ttl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PresignedURL"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid S3 key",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/media/upload": {
            "get": {
                "description": "Generate a presigned POST URL for uploading media files.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Get presigned POST URL for media upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media upload intent",
                        "name": "intent",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filename to be uploaded",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Content type of the file",
                        "name": "content_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.POSTUploadPolicy"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.PresignedURL": {
            "type": "object",
            "properties": {
                "presigned_url": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                }
            }
        },
...
        "model.APIResponse": {
            "type": "object",
            "required": [
                "request_id",
                "status"
            ],
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "status": {
                    "type": "integer",
                    "maximum": 599,
                    "minimum": 100
                }
            }
        },
...
        "model.POSTUploadPolicy": {
            "type": "object",
            "required": [
                "formFields",
                "prefix",
                "url"
            ],
            "properties": {
                "formFields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "prefix": {
                    "type": "string",
                    "example": "avatars/"
                },
                "url": {
                    "type": "string",
                    "example": "https://s3.planetleague.co.uk/bucket/key"
                }
            }
        }
    }
}